Write speeds when processing.
Not from files produced while processing.
