+++ cvts/tasks/__init__.py
@@ -2,4 +2,3 @@
 
 from ._valhalla import ListRawFiles, MatchToNetwork
 from ._regiondensity import RegionCounts, RasterCounts, SourceDestinationCounts
-from ._speed import AverageSpeed, TripsToDB
